+++ example/a.ts
@@ -1,6 +1,11 @@
-import { Elysia } from '../src'
+import { Elysia, t } from '../src'
 import homepage from './index.html'
 
 new Elysia()
-	.get('/', homepage)
+	.onError(({ status }) => {
+		console.log(status(400, 2))
+	})
+	.get('/', () => {
+		throw new Error('This is an error')
+	})
 	.listen(3000, (server) => console.log(`Running on ${server.url}`))